Port from qt3 to qt4.  Removes a load of warnings.
svn path=/trunk/KDE/kdebase/workspace/; revision=517609

+++ gui/SensorDisplayLib/FancyPlotter.cc
@@ -187,8 +187,8 @@ void FancyPlotter::applyStyle()
   mPlotter->setHorizontalLinesColor( KSGRD::Style->secondForegroundColor() );
   mPlotter->setBackgroundColor( KSGRD::Style->backgroundColor() );
   mPlotter->setFontSize( KSGRD::Style->fontSize() );
-  for ( unsigned int i = 0; i < mPlotter->beamColors().count() &&
-        i < KSGRD::Style->numSensorColors(); ++i )
+  for ( int i = 0; i < mPlotter->beamColors().count() &&
+        (unsigned int)i < KSGRD::Style->numSensorColors(); ++i )
     mPlotter->beamColors()[ i ] = KSGRD::Style->sensorColor( i );
 
   mPlotter->update();
